refactor(types): Move Config struct into types package
