+++ types/types.go
@@ -5,3 +5,15 @@ type Movie struct {
 	ID    string `json:"id"`
 	Title string `json:"title"`
 }
+
+type Config struct {
+	Database Database
+	Env      string
+}
+
+type Database struct {
+	Username string
+	Password string
+	Host     string
+	URL      string
+}
